mcp-servers/mock-oidc/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
